refactored endpoints and spout
