+++ replacer/replacer.go
@@ -1,0 +1,5 @@
+package replacer
+
+type Replacer interface {
+	// TODO
+}
