+++ rest/endpoints.go
@@ -1,4 +1,4 @@
-package main
+package rest
 
 import (
 	"encoding/json"
